requirements.txt
torch>=2.0.0
transformers>=4.30.0
numpy>=1.21.0
scipy>=1.7.0
matplotlib>=3.5.0
seaborn>=0.11.0
wandb>=0.15.0
biotite>=0.37.0
biopython>=1.79
einops>=0.6.0
timm>=0.9.0
pytest>=7.0.0
black>=22.0.0
flake8>=4.0.0
pyyaml>=6.0.0
scikit-learn>=1.0.0
pandas>=1.3.0
tqdm>=4.60.0